+++ functions/Get-GTPIMRoleReport.ps1
@@ -45,19 +45,18 @@ function Get-GTPIMRoleReport
 
     begin
     {
-        # 1. Scope Check
+        # 1. Connection Initialization
         $requiredScopes = @('RoleManagement.Read.Directory', 'User.Read.All', 'Group.Read.All')
-        
-        if (-not (Test-GTGraphScopes -RequiredScopes $requiredScopes -Reconnect -Quiet))
-        {
-            Write-Error "Failed to acquire required permissions ($($requiredScopes -join ', ')). Aborting."
-            return
-        }
-
-        # 2. Connection Initialization
         if (-not (Initialize-GTGraphConnection -Scopes $requiredScopes -NewSession:$NewSession))
         {
             Write-Error "Failed to initialize session."
+            return
+        }
+
+        # 2. Scope Validation
+        if (-not (Test-GTGraphScopes -RequiredScopes $requiredScopes -Quiet))
+        {
+            Write-Error "Failed to acquire required permissions ($($requiredScopes -join ', ')). Aborting."
             return
         }