+++ functions/Invoke-AuditLogQuery.ps1
@@ -73,7 +73,8 @@ function Invoke-AuditLogQuery
     [CmdletBinding()]
     param(
         [Parameter(Mandatory = $false)]
-        [string[]]$RequieredScopes = @('AuditLogsQuery-CRM.Read.All', 'AuditLogsQuery-Endpoint.Read.All', 'AuditLogsQuery-Exchange.Read.All', 'AuditLogsQuery-OneDrive.Read.All', 'AuditLogsQuery-SharePoint.Read.All', 'AuditLogsQuery.Read.All'),
+        [Alias('RequieredScopes', 'Scope', 'Scopes')]
+        [string[]]$RequiredScopes = @('AuditLogsQuery-CRM.Read.All', 'AuditLogsQuery-Endpoint.Read.All', 'AuditLogsQuery-Exchange.Read.All', 'AuditLogsQuery-OneDrive.Read.All', 'AuditLogsQuery-SharePoint.Read.All', 'AuditLogsQuery.Read.All'),
 
         [Parameter(Mandatory = $false)]
         [switch]$NewSession,
@@ -151,13 +152,12 @@ function Invoke-AuditLogQuery
 
 
         # Connect to Microsoft Graph
-        if ($NewSession)
+        if (-not (Initialize-GTGraphConnection -Scopes $RequiredScopes -NewSession:$NewSession))
         {
-            Write-PSFMessage -Level 'Verbose' -Message 'Closing existing Microsoft Graph session.'
-            Disconnect-GTGraph
+            throw "Graph connection failed: Could not initialize session."
         }
 
-        $session = Test-GTGraphScopes -RequiredScopes $RequiredScopes -Reconnect -Quiet
+        $session = Test-GTGraphScopes -RequiredScopes $RequiredScopes -Quiet
         if (-not $session)
         {
             throw "Graph connection failed: Required scopes not available"